Renaming

+++ src/main/java/com/vikhani/animfeigner/dtos/AnimalResultDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class AnimalResultDto {
-    private short httpStatus;
+    private int httpStatus;
     private String animalNames;
 }
